pyproject.toml
[tool.poetry]
name = "test-relations"
version = "0.1.0"
description = ""
authors = ["Saveliy <s.lekhtin@innopolis.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
ruff = "^0.1.6"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
alembic = "^1.12.1"

[tool.ruff]
# Set the maximum line length to 120.
line-length = 120



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
